FSA-5722: Refactored tests structure
