+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/bronze-cancel-change.e2e-spec.ts
