+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/bronze-change-coverage.e2e-spec.ts
