+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/gold-add-driver-remove-coverage.e2e-spec.ts
