+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/gold-fnol.e2e-spec.ts
