+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/silver-change-mileage.e2e-spec.ts
