+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/silver-coupons.e2e-spec.ts
