+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/autoe2e/silver-referred-quote.e2e-spec.ts
