+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/banking/credit-card-reload.e2e-spec.ts
