+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/banking/credit-card.e2e-spec.ts
