+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/banking/fixed-term-deposit.e2e-spec.ts
