+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/common/asm.e2e-spec.ts
