+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/common/consent-management.e2e-spec.ts
