+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/common/homepage.e2e-spec.ts
