+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/common/inbox.e2e-spec.ts
