+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/common/search.e2e-spec.ts
