+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/group-policy/group-policy-root-admin.e2e-spec.ts
