+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/insurance/life.e2e-spec.ts
