+++ projects/fsastorefront-e2e-cypress/cypress/integration/regression/insurance/retrieve-quote.e2e-spec.ts
