+++ projects/fsastorefront-e2e-cypress/cypress/integration/smoke/current-account.e2e-spec.ts
@@ -1,9 +1,9 @@
-import * as banking from '../../../helpers/checkout/banking/checkout-banking';
-import * as checkout from '../../../helpers/checkout/checkout-steps';
-import * as currentAccount from '../../../helpers/checkout/banking/current-account';
-import * as register from '../../../helpers/register';
-import { registrationUser } from '../../../sample-data/users';
-import * as userIdentification from '../../../helpers/checkout/banking/user-identification';
+import * as banking from '../../helpers/checkout/banking/checkout-banking';
+import * as checkout from '../../helpers/checkout/checkout-steps';
+import * as currentAccount from '../../helpers/checkout/banking/current-account';
+import * as register from '../../helpers/register';
+import { registrationUser } from '../../sample-data/users';
+import * as userIdentification from '../../helpers/checkout/banking/user-identification';
 
 context('Current Account Checkout', () => {
   before(() => {
