+++ projects/fsastorefront-e2e-cypress/cypress/integration/smoke/loan.e2e-spec.ts
@@ -1,10 +1,10 @@
-import * as register from '../../../helpers/register';
-import { registrationUser } from '../../../sample-data/users';
-import * as checkout from '../../../helpers/checkout/checkout-steps';
-import * as banking from '../../../helpers/checkout/banking/checkout-banking';
-import * as loan from '../../../helpers/checkout/banking/loan';
-import * as userIdentification from '../../../helpers/checkout/banking/user-identification';
-import * as policies from '../../../helpers/my-account/policies';
+import * as register from '../../helpers/register';
+import { registrationUser } from '../../sample-data/users';
+import * as checkout from '../../helpers/checkout/checkout-steps';
+import * as banking from '../../helpers/checkout/banking/checkout-banking';
+import * as loan from '../../helpers/checkout/banking/loan';
+import * as userIdentification from '../../helpers/checkout/banking/user-identification';
+import * as policies from '../../helpers/my-account/policies';
 
 context('Loan Checkout', () => {
   before(() => {
