+++ projects/fsastorefront-e2e-cypress/cypress/integration/smoke/product-category-page.e2e-spec.ts
@@ -1,5 +1,5 @@
-import * as productCategory from '../../../helpers/product-category-page';
-import { insuranceButtons, bankingButtons } from '../../../helpers/constants';
+import * as productCategory from '../../helpers/product-category-page';
+import { insuranceButtons, bankingButtons } from '../../helpers/constants';
 
 context('ProductCategoryPage', () => {
   before(() => {
