+++ projects/fsastorefront-e2e-cypress/package.json
@@ -5,11 +5,12 @@
   "scripts": {
     "cy:open": "cypress open",
     "cy:open:local": "cypress open --config-file cypress.local.json",
-    "cy:run": "cypress run --spec cypress/integration/checkout/**/*  --browser=chrome",
-    "cy:run:insurance": "cypress run --spec cypress/integration/checkout/insurance/**/*",
-    "cy:run:banking": "cypress run --spec cypress/integration/checkout/banking/**/*",
-    "cy:run:auto": "cypress run --spec cypress/integration/checkout/autoe2e/**/*",
-    "cy:run:common": "cypress run --spec cypress/integration/checkout/common/**/*",
+    "cy:run": "cypress run --spec cypress/integration/regression/**/*  --browser=chrome",
+    "cy:run:smoke": "cypress run --spec cypress/integration/smoke/**/*  --browser=chrome",
+    "cy:run:insurance": "cypress run --spec cypress/integration/regression/insurance/**/*",
+    "cy:run:banking": "cypress run --spec cypress/integration/regression/banking/**/*",
+    "cy:run:auto": "cypress run --spec cypress/integration/regression/autoe2e/**/*",
+    "cy:run:common": "cypress run --spec cypress/integration/regression/common/**/*",
     "cy:run:local": "cypress run --config-file cypress.local.json",
     "cy:cli": "cypress run",
     "cy:run:mobile": "cypress run --config numTestsKeptInMemory=1 --spec cypress/integration/mobile/**/*",
